refactored module linter
