+++ utils/mlinter/__main__.py
@@ -1,0 +1,4 @@
+from .mlinter import main
+
+
+raise SystemExit(main())
